refactor(cli): extract config path boundary
